diff2flow/requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0

# Diffusion / Generative Models
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Parameter-Efficient Finetuning
peft>=0.7.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data & Config
omegaconf>=2.3.0
hydra-core>=1.3.0
datasets>=2.16.0
webdataset>=0.2.48

# Logging & Metrics
wandb>=0.16.0
tensorboard>=2.15.0
torchmetrics>=1.2.0

# Evaluation
scipy>=1.11.0
scikit-image>=0.22.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
einops>=0.7.0
matplotlib>=3.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# HuggingFace Hub
huggingface-hub>=0.20.0
